Adding note about 3rd party embeds blog
I took the part about 3rd party embeds largely from the Medium privacy policy. https://medium.com/policy/medium-privacy-policy-f03bf92035c9 I am not going into detail what data is shared with these 3rd parties etc. I hope keeping it this vague is legally fine.

The second question is about whether or not we need to add anything about the links we have to social media on the blog, on the info pages, on the index page and in other places. As far as I can tell we don't need to add anything to our privacy policy as long as these are just simple links and not plugins or the like. But I was not able to find any final confirmation of that.