We don't collect IMEI ourselves
Turns out that we are not collecting the IMEI through our iOS and Android app ourselves. That is good we don't need that unique identifier but the policy stated wrongly that we do. Now fixed. This ID is still collected by hockeyapp for both apps and the facebook SDK for Android.